@@ -22,7 +22,7 @@
 		"issues": "https://github.com/kdyby/clock/issues"
 	},
 	"require": {
-		"php": "^5.6 || ^7.0",
+		"php": "^8.0",
 		"kdyby/strict-objects": "^1.0"
 	},
 	"require-dev": {
